# perf(install): only read initialized file if we care about the tags